# apps/api/openapi/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sessions": {
      "post": {
        "operationId": "InterviewController_createSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInterviewSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Create a single-question interview session from a generated note question",
        "tags": [
          "interview"
        ]
      },
      "get": {
        "operationId": "InterviewController_findSessions",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewSessionResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "List interview sessions for the current user",
        "tags": [
          "interview"
        ]
      }
    },
    "/sessions/{sessionId}": {
      "get": {
        "operationId": "InterviewController_findSessionById",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Get an interview session by id",
        "tags": [
          "interview"
        ]
      },
      "patch": {
        "operationId": "InterviewController_updateSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitInterviewAnswerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Submit or replace the answer for a single-question interview session",
        "tags": [
          "interview"
        ]
      },
      "delete": {
        "operationId": "InterviewController_deleteSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Delete an interview session",
        "tags": [
          "interview"
        ]
      }
    },
    "/sessions/{sessionId}/answer": {
      "post": {
        "operationId": "InterviewController_answerQuestion",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitInterviewAnswerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Answer the generated question in a single-question interview session",
        "tags": [
          "interview"
        ]
      }
    },
    "/sessions/multi-turn": {
      "post": {
        "operationId": "InterviewController_startMultiTurnSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartMultiTurnSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Start a resumable multi-turn interview session",
        "tags": [
          "interview"
        ]
      }
    },
    "/sessions/{sessionId}/turns": {
      "post": {
        "operationId": "InterviewController_submitTurn",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTurnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitTurnResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Submit one candidate turn in a multi-turn interview session",
        "tags": [
          "interview"
        ]
      },
      "get": {
        "operationId": "InterviewController_getTurns",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InterviewTurnDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Get all turns for a multi-turn interview session",
        "tags": [
          "interview"
        ]
      }
    },
    "/sessions/{sessionId}/end": {
      "post": {
        "operationId": "InterviewController_endSession",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewSessionResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "End a multi-turn interview session and trigger summary/evaluation work",
        "tags": [
          "interview"
        ]
      }
    },
    "/sessions/{sessionId}/evaluation": {
      "get": {
        "operationId": "InterviewController_getEvaluation",
        "parameters": [
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewEvaluationDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Get the evaluation for an interview session",
        "tags": [
          "interview"
        ]
      }
    },
...
    "/english-notes": {
      "get": {
        "operationId": "EnglishNotesController_findEnglishNotes",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EnglishNoteResponseDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "List english improvement notes",
        "tags": [
          "english-notes"
        ]
      },
      "post": {
        "operationId": "EnglishNotesController_createEnglishNote",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEnglishNoteDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnglishNoteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Create an english improvement note",
        "tags": [
          "english-notes"
        ]
      }
    },
    "/english-notes/{englishNoteId}/status": {
      "patch": {
        "operationId": "EnglishNotesController_updateEnglishNoteStatus",
        "parameters": [
          {
            "name": "englishNoteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEnglishNoteStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnglishNoteResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "interviewos_session": []
          }
        ],
        "summary": "Update english note status",
        "tags": [
          "english-notes"
        ]
      }
    },
...
  },
  "info": {
    "title": "InterviewOS API",
    "description": "REST API for InterviewOS local development and integration testing.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "interviewos_session"
      }
    },
    "schemas": {
...
      "InterviewEvaluationDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "overallScore": {
            "type": "number",
            "nullable": true
          },
          "summary": {
            "type": "string",
            "nullable": true
          },
          "confidence": {
            "type": "number",
            "nullable": true
          },
          "dimensionScores": {
            "type": "object",
            "additionalProperties": true
          },
          "starScores": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "designScores": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "rubricScores": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "weaknesses": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "recommendations": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "strengths": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "improvements": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "coachingNotes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "weakConcepts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sessionId",
          "status",
          "overallScore",
          "summary",
          "confidence",
          "dimensionScores",
          "starScores",
          "designScores",
          "rubricScores",
          "evidence",
          "weaknesses",
          "recommendations",
          "strengths",
          "improvements",
          "coachingNotes",
          "weakConcepts",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "InterviewSessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "TECHNICAL",
              "BEHAVIORAL",
              "SYSTEM_DESIGN",
              "MIXED"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "REVIEWING",
              "NEEDS_PRACTICE",
              "INTERVIEW_READY",
              "MASTERED",
              "PUBLISHED",
              "ARCHIVED"
            ]
          },
          "noteId": {
            "type": "string",
            "nullable": true
          },
          "sourceQuestionId": {
            "type": "string",
            "nullable": true
          },
          "overrideRole": {
            "type": "string",
            "nullable": true
          },
          "overrideLevel": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR",
              "STAFF",
              "PRINCIPAL"
            ],
            "nullable": true
          },
          "overrideStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "overrideGoals": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "overrideEnglishLevel": {
            "type": "string",
            "enum": [
              "BEGINNER",
              "ELEMENTARY",
              "INTERMEDIATE",
              "UPPER_INTERMEDIATE",
              "ADVANCED",
              "NATIVE"
            ],
            "nullable": true
          },
          "preferredOutputStyle": {
            "type": "string",
            "nullable": true
          },
          "mode": {
            "type": "string"
          },
          "companyModeId": {
            "type": "string",
            "nullable": true
          },
          "parentSessionId": {
            "type": "string",
            "nullable": true
          },
          "version": {
            "type": "number"
          },
          "maxTurns": {
            "type": "number"
          },
          "currentTurnNum": {
            "type": "number"
          },
          "lastActivityAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "contextSnapshot": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewQuestionDto"
            }
          },
          "note": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true
          },
          "companyMode": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/CompanyModeReferenceDto"
              }
            ]
          },
          "evaluation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewEvaluationDto"
              }
            ]
          },
          "summary": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewSummaryDto"
              }
            ]
          },
          "readinessImpact": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewReadinessImpactDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "type",
          "status",
          "noteId",
          "sourceQuestionId",
          "overrideRole",
          "overrideStack",
          "overrideGoals",
          "preferredOutputStyle",
          "companyModeId",
          "parentSessionId",
          "lastActivityAt",
          "contextSnapshot",
          "startedAt",
          "endedAt",
          "createdAt",
          "updatedAt",
          "questions",
          "note"
        ]
      },
      "CreateInterviewSessionDto": {
        "type": "object",
        "properties": {
          "generatedQuestionId": {
            "type": "string"
          }
        },
        "required": [
          "generatedQuestionId"
        ]
      },
...
      "SubmitInterviewAnswerDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string"
          },
          "advancedSettings": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewAdvancedSettingsDto"
              }
            ]
          }
        },
        "required": [
          "answer"
        ]
      },
      "StartMultiTurnSessionDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "TECHNICAL",
              "BEHAVIORAL",
              "SYSTEM_DESIGN",
              "MIXED"
            ]
          },
          "companyModeSlug": {
            "type": "string"
          },
          "noteId": {
            "type": "string"
          },
          "overrideRole": {
            "type": "string"
          },
          "overrideLevel": {
            "type": "string",
            "enum": [
              "JUNIOR",
              "MID",
              "SENIOR",
              "STAFF",
              "PRINCIPAL"
            ]
          },
          "overrideStack": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "maxTurns": {
            "type": "number",
            "minimum": 1,
            "maximum": 30
          }
        },
        "required": [
          "type"
        ]
      },
...
      "InterviewTurnDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionId": {
            "type": "string"
          },
          "turnNumber": {
            "type": "number"
          },
          "role": {
            "type": "string",
            "enum": [
              "INTERVIEWER",
              "CANDIDATE"
            ]
          },
          "content": {
            "type": "string"
          },
          "decision": {
            "type": "string",
            "enum": [
              "FOLLOW_UP",
              "DEEP_DIVE",
              "CLARIFY",
              "CHALLENGE",
              "ESCALATE",
              "ADVANCE",
              "WRAP_UP",
              "EVALUATE"
            ],
            "nullable": true
          },
          "topicTags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "reasoning": {
            "type": "string",
            "nullable": true
          },
          "references": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InterviewTurnReferenceDto"
            }
          },
          "evaluation": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewTurnEvaluationDto"
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "sessionId",
          "turnNumber",
          "role",
          "content",
          "topicTags",
          "reasoning",
          "createdAt"
        ]
      },
      "SubmitTurnResponseDto": {
        "type": "object",
        "properties": {
          "candidateTurn": {
            "$ref": "#/components/schemas/InterviewTurnDto"
          },
          "interviewerTurn": {
            "nullable": true,
            "type": "object",
            "allOf": [
              {
                "$ref": "#/components/schemas/InterviewTurnDto"
              }
            ]
          },
          "decision": {
            "type": "string",
            "enum": [
              "FOLLOW_UP",
              "DEEP_DIVE",
              "CLARIFY",
              "CHALLENGE",
              "ESCALATE",
              "ADVANCE",
              "WRAP_UP",
              "EVALUATE"
            ]
          },
          "turnNumber": {
            "type": "number"
          },
          "isComplete": {
            "type": "boolean"
          }
        },
        "required": [
          "candidateTurn",
          "decision",
          "turnNumber",
          "isComplete"
        ]
      },
      "SubmitTurnDto": {
        "type": "object",
        "properties": {
          "answer": {
            "type": "string"
          }
        },
        "required": [
          "answer"
        ]
      },
...
      "EnglishNoteResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "answerId": {
            "type": "string"
          },
          "originalSentence": {
            "type": "string"
          },
          "correctedSentence": {
            "type": "string"
          },
          "naturalVersion": {
            "type": "string"
          },
          "explanation": {
            "type": "string"
          },
          "grammarTopic": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "NEEDS_PRACTICE",
              "REVIEWING",
              "IMPROVED",
              "MASTERED"
            ]
          },
          "recommendedStudyTopics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "practicePatterns": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "userId",
          "answerId",
          "originalSentence",
          "correctedSentence",
          "naturalVersion",
          "explanation",
          "grammarTopic",
          "status",
          "recommendedStudyTopics",
          "practicePatterns",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateEnglishNoteStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "NEEDS_PRACTICE",
              "REVIEWING",
              "IMPROVED",
              "MASTERED"
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}